requirements.txt
streamlit==1.39.0
pandas==2.2.3
google-generativeai==0.8.3
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
apscheduler==3.10.4
python-dotenv==1.0.1
PyPDF2==3.0.1
lxml==5.3.0
streamlit
pandas
gspread
oauth2client
google-generativeai
requests
beautifulsoup4
feedparser
apscheduler
selenium
webdriver-manager